requirements.txt
# Core Framework & Server
fastapi==0.100.0
uvicorn[standard]>=0.30.1,<0.31.0
pydantic>=2.7.3,<2.8.0
pydantic-settings>=2.2.1,<2.3.0

# Database (Tortoise ORM)
tortoise-orm>=0.21.2,<0.22.0
asyncpg>=0.29.0,<0.30.0
psycopg2-binary>=2.9.9,<3.0.0

# Database Migration
aerich==0.8.2

# Authentication & Security
passlib[bcrypt]>=1.7.4,<1.8.0
bcrypt==4.0.1
python-jose[cryptography]>=3.3.0,<3.4.0
python-multipart>=0.0.7,<0.0.8

# Logging
loguru>=0.7.2,<0.8.0

# Utilities
python-dotenv>=1.0.0,<1.1.0
email-validator>=2.1.0,<2.2.0

# Testing
pytest>=7.4.0,<7.5.0
pytest-asyncio>=0.21.1,<0.22.0
httpx>=0.24.1,<0.25.0
pytest-cov>=4.1.0,<4.2.0

aerich==0.8.2